+++ lab1/logarithm.cpp
@@ -2,13 +2,13 @@
 #include "logarithm.h"
 #include <cmath>
 
-double CountLogarithm(double x, int n)
+double CountLogarithm(double x, long long int n)
 {
     CheckInput(x, n);
 
     double result = 0;
 
-    for (int i = 1; i <= n; i++)
+    for (long long int i = 1; i <= n; i++)
     {
         result += pow(-1, i + 1) * x / i;
         x *= x;
